+++ crates/consensus/src/transaction/mod.rs
@@ -1,13 +1,13 @@
-//! Tramsaction types for Optimism.
+//! Transaction types for Optimism.
 
 mod deposit;
 pub use deposit::{DepositTransaction, TxDeposit};
 
 mod tx_type;
-pub use tx_type::{DEPOSIT_TX_TYPE_ID, OpTxType};
+pub use tx_type::DEPOSIT_TX_TYPE_ID;
 
 mod envelope;
-pub use envelope::OpTxEnvelope;
+pub use envelope::{OpTransaction, OpTxEnvelope, OpTxType};
 
 #[cfg(all(feature = "serde", feature = "serde-bincode-compat"))]
 pub use envelope::serde_bincode_compat as envelope_serde_bincode_compat;
@@ -20,6 +20,9 @@ pub use pooled::OpPooledTransaction;
 
 #[cfg(feature = "serde")]
 pub use deposit::serde_deposit_tx_rpc;
+
+mod meta;
+pub use meta::{OpDepositInfo, OpTransactionInfo};
 
 /// Bincode-compatible serde implementations for transaction types.
 #[cfg(all(feature = "serde", feature = "serde-bincode-compat"))]